Revamped GUI interface and command system
Edited the basic layout to have the controls in a tab group and the drone selector + others outside of the tab.

Added a frame continuing information about the currently selected drone

Values in the controls now update to the current values of the selected drone when a new drone is selected.

Values in the info section update when switching to a new drone or when the update button is clicked.

Fixed issue where ground control would throw an error when the window gets closed

Added a query command to the list of commands. This lets a client retrieve certain values stored in the drone controller

Changes the SendCommand function to use the drone’s name instead of the drone number